+++ src/main/java/ru/practicum/shareit/user/storage/UserInMemoryStorage.java
@@ -9,10 +9,10 @@ import ru.practicum.shareit.user.User;
 
 @Component
 public class UserInMemoryStorage implements UserStorage {
-  private final Map<Integer, User> idToUser = new HashMap<>();
-  private Integer id = 1;
+  private final Map<Long, User> idToUser = new HashMap<>();
+  private Long id = 1L;
 
-  private Integer getNextId() {
+  private Long getNextId() {
     return id++;
   }
 
@@ -29,12 +29,12 @@ public class UserInMemoryStorage implements UserStorage {
   }
 
   @Override
-  public void delete(Integer id) {
+  public void delete(Long id) {
     idToUser.remove(id);
   }
 
   @Override
-  public User findById(Integer id) {
+  public User findById(Long id) {
     return idToUser.get(id);
   }
